MAINT: Moved TimeStepManager to time module
