+++ src/porepy/__init__.py
@@ -166,7 +166,7 @@ from porepy.numerics.ad.ad_utils import (
 )
 
 # Time stepping control
-from porepy.numerics.time_step_control import TimeManager
+from porepy.time.time_step_control import TimeManager
 
 from porepy import models
 from porepy.models.abstract_equations import (
